+++ src/vue/modules/assets/asset-explorer/components/assets-renderer.vue
@@ -8,7 +8,7 @@
           </template>
 
           <asset-form
-            :collector="collector"
+            :former="former"
             @submitted="onAssetUpdate()"
           />
         </template>
@@ -94,7 +94,7 @@ import { vuexTypes } from '@/vuex'
 
 import { ErrorHandler } from '@/js/helpers/error-handler'
 import UpdateList from '@/vue/mixins/update-list.mixin'
-import { AssetCollector } from '@/js/collectors/AssetCollector'
+import { AssetFormer } from '@/js/formers/AssetFormer'
 
 export default {
   name: 'assets-renderer',
@@ -139,7 +139,7 @@ export default {
     isDrawerShown: false,
     isAssetFormShown: false,
     selectedAsset: {},
-    collector: new AssetCollector('update'),
+    former: new AssetFormer('update'),
     itemsPerSkeletonLoader: 3,
   }),
 
@@ -209,7 +209,7 @@ export default {
     },
 
     showUpdateForm () {
-      this.collector.from(this.selectedAsset)
+      this.former.populate(this.selectedAsset)
       this.isAssetFormShown = true
     },