+++ src/vue/modules/assets/balance-explorer/index.vue
@@ -8,7 +8,7 @@
           </template>
 
           <asset-form
-            :collector="collector"
+            :former="former"
             @submitted="onAssetUpdate()"
           />
         </template>
@@ -85,7 +85,7 @@ import { vuexTypes } from '@/vuex'
 
 import { ErrorHandler } from '@/js/helpers/error-handler'
 import UpdateList from '@/vue/mixins/update-list.mixin'
-import { AssetCollector } from '@/js/collectors/AssetCollector'
+import { AssetFormer } from '@/js/formers/AssetFormer'
 
 export default {
   name: 'balance-explorer',
@@ -108,7 +108,7 @@ export default {
     selectedBalance: {
       asset: {},
     },
-    collector: new AssetCollector('update'),
+    former: new AssetFormer('update'),
     itemsPerSkeletonLoader: 3,
   }),
 
@@ -153,7 +153,7 @@ export default {
     },
 
     showUpdateForm () {
-      this.collector.from(this.selectedBalance.asset)
+      this.former.populate(this.selectedBalance.asset)
       this.isAssetFormShown = true
     },