+++ src/vue/modules/requests/update-asset-requests/index.vue
@@ -7,7 +7,7 @@
             {{ 'update-asset-requests.update-asset-title' | globalize }}
           </template>
           <asset-form
-            :collector="collector"
+            :former="former"
             @submitted="onRequestUpdate()"
           />
         </template>
@@ -59,7 +59,7 @@ import { types } from './store/types'
 
 import { ErrorHandler } from '@/js/helpers/error-handler'
 import UpdateList from '@/vue/mixins/update-list.mixin'
-import { AssetCollector } from '@/js/collectors/AssetCollector'
+import { AssetFormer } from '@/js/formers/AssetFormer'
 
 export default {
   name: 'update-asset-requests-module',
@@ -79,7 +79,7 @@ export default {
     isDrawerShown: false,
     isAssetFormShown: false,
     selectedRequest: {},
-    collector: new AssetCollector('update'),
+    former: new AssetFormer('update'),
     firstPageLoader: () => {},
   }),
 
@@ -131,7 +131,7 @@ export default {
     },
 
     showUpdateForm () {
-      this.collector.from(this.selectedRequest)
+      this.former.populate(this.selectedRequest)
       this.isAssetFormShown = true
     },